+++ CHANGELOG.md
@@ -3,10 +3,11 @@
 ## **6.1.0** (in progress)
 
 ### Features added:
-* Add accessibiity testing to automated tests in ui-data-import (UIDATIMP-1372)
+* Add accessibility testing to automated tests in ui-data-import (UIDATIMP-1372)
 
 ### Bugs fixed:
 * Order field mapping profile: Fix the UI and error message for the fund "Value" field (UIDATIMP-1387)
+* Field mapping profile Invoice Edit screen: Accounting code field is disabled when Vendor name is selected (UIDATIMP-1394)
 * Order field mapping profile: Add info icon to the Purchase order status fieldPurchase order status a required field (UIDATIMP-1397)
 * Order field mapping profile: Inactivate Create Inventory fields when Order status is Open (UIDATIMP-1402)